refactor(backend): move project membership to its own service
organization.py held two unrelated concerns: seeding a new org's demo
data, and project membership writes. The membership half moves to
services/project_membership.py, which already owned the read side
(list_other_member_projects), so both sides now live together.

Moves the three exception classes, the default_project repair helpers,
_bust_permission_cache and the enroll/unenroll functions. load_initial_data
calls enroll_user_in_project, so organization.py now imports it; no cycle,
project_membership depends only on models and scope.

Also drops a duplicate _logger assignment that the move made visible --
organization.py defined it twice, the second silently rebinding the first.

Pure relocation: the two oversized functions stay in organization.py, so
its complexity per-file-ignores entry is unchanged.
